move fetch_faqs server fn into accordion_menu.rs - clean up server.rs
