--- src/server.rs
@@ -1,1 +1,0 @@
-pub mod api;
