--- src/server/api.rs
@@ -1,1 +1,0 @@
-pub mod fetch_faqs;
